Properly OptiPNG'd.
A first run of 'optipng -o7 -zm1-9 -nx *.png' is used, then followed by a run of 'optipng -o7 -zm1-9 *.png'. The -nx flag specifies not to do any reductions, meaning converting RGBA to just a <256 color pallete for example. In some cases, a color pallete can actually increase the size of the image, which OptiPNG will pick up on and not output a bigger image. However, if you run over it with the -nx flag, you can still possibly get a reduction through the better compression as usual, AND even if a reduction-thingy results in a smaller output than the original, you can still potentially lose out on just a smaller image without the reduction stuff! I don't know if this explanation really makes sense, but um... yeah.